requirements.txt
# ============================================
# Dependencias del Sistema de Trading Autónomo
# ============================================

# ===== Proveedores de IA =====
openai>=1.0.0                    # Para DeepSeek y OpenAI
google-generativeai>=0.3.0       # Para Google Gemini

# ===== Exchanges y Brokers =====
ccxt[async]>=4.2.0               # Conexión con exchanges + soporte async/WebSockets
aiohttp>=3.9.0                   # HTTP async para WebSockets
# ib_insync>=0.9.86              # Interactive Brokers (descomentar si se usa)

# ===== Análisis Técnico =====
pandas>=2.0.0                    # Manipulación de datos temporales
numpy>=1.24.0                    # Cálculos numéricos
ta>=0.11.0                       # Indicadores técnicos (compatible con todas las plataformas)
# pandas-ta>=0.3.14b             # Alternativa local (pip install pandas-ta)
# ta-lib>=0.4.28                 # Requiere librería C de TA-Lib

# ===== Configuración y Variables de Entorno =====
python-dotenv>=1.0.0             # Gestión de variables de entorno
pyyaml>=6.0                      # Parsing de archivos YAML

# ===== Notificaciones (Opcional) =====
python-telegram-bot>=20.0        # Notificaciones vía Telegram
requests>=2.31.0                 # Para APIs HTTP

# ===== Base de Datos (Opcional) =====
influxdb-client>=1.38.0          # InfluxDB para métricas y backtesting

# ===== Utilidades =====
python-dateutil>=2.8.2           # Manejo de fechas
pytz>=2023.3                     # Zonas horarias

# ===== Validación de Datos =====
pydantic>=2.0.0                  # Validación robusta de respuestas IA

# ===== Testing y Calidad de Código =====
pytest>=7.4.0                    # Testing
pytest-cov>=4.1.0                # Cobertura de tests
black>=23.0.0                    # Formateo de código
flake8>=6.1.0                    # Linting

# ===== Logging Avanzado (Opcional) =====
colorlog>=6.7.0                  # Logs con colores en terminal

# ===== Backtesting (Opcional - descomentar si se necesita) =====
# backtrader>=1.9.78.123         # Framework de backtesting
# vectorbt>=0.25.0               # Backtesting vectorizado de alta velocidad

# ===== Visualización (Opcional - descomentar si se necesita) =====
# matplotlib>=3.7.0              # Gráficos
# plotly>=5.17.0                 # Gráficos interactivos
# mplfinance>=0.12.10b           # Gráficos financieros

# ===== Monitoreo de Performance (Opcional) =====
psutil>=5.9.0                    # Monitoreo de recursos del sistema
